add json-types crate
