+++ crates/fiber-json-types/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "fiber-json-types"
+version = "0.7.1"
+edition = "2021"
+description = "RPC JSON request/response types for the Fiber Network"
+
+[dependencies]
+# Serialization
+serde = { version = "1.0", features = ["derive"] }
+serde_with = { version = "3.7", features = ["macros"] }
+
+# Core domain types
+fiber-types = { path = "../fiber-types", features = ["watchtower"] }
+
+# CKB types
+ckb-jsonrpc-types = "1"
+ckb-types = "1"
+
+# Cryptography (for PublicKey in invoice Attribute)
+secp256k1 = { version = "0.30.0", features = ["serde"] }
+
+# Network (for MultiAddr in peer/graph types)
+tentacle-multiaddr = "0.3"
+
+[features]
+default = []
+cch = ["fiber-types/cch"]
+
+[lints.clippy]
+expect-fun-call = "allow"
+needless-return = "allow"
+upper-case-acronyms = "allow"
