+++ crates/fiber-lib/src/rpc/cch.rs
@@ -1,68 +1,11 @@
-use crate::{cch::CchMessage, invoice::Currency};
-use fiber_types::{CchInvoice, CchOrder, CchOrderStatus, Hash256, U128Hex, U64Hex};
+use crate::cch::CchMessage;
 use jsonrpsee::{
     proc_macros::rpc,
     types::{error::CALL_EXECUTION_FAILED_CODE, ErrorObjectOwned},
 };
 use ractor::{call_t, ActorRef};
-use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
 
-#[derive(Serialize, Deserialize)]
-pub struct SendBTCParams {
-    /// Payment request string for the BTC Lightning payee.
-    ///
-    /// The invoice should not be expired soon. The remaining expiry time should be greater than the CCH config
-    /// `min_incoming_invoice_expiry_delta_seconds`.
-    pub btc_pay_req: String,
-    /// Request currency
-    pub currency: Currency,
-}
-
-#[serde_as]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CchOrderResponse {
-    /// Seconds since epoch when the order is created
-    #[serde_as(as = "U64Hex")]
-    pub timestamp: u64,
-    /// Relative expiry time in seconds from `created_at` that the order expires
-    #[serde_as(as = "U64Hex")]
-    pub expiry_delta_seconds: u64,
-
-    /// Wrapped BTC type script
-    pub wrapped_btc_type_script: ckb_jsonrpc_types::Script,
-
-    /// Generated invoice for the incoming payment
-    pub incoming_invoice: CchInvoice,
-    /// The final payee to accept the payment. It has the different network with incoming invoice.
-    pub outgoing_pay_req: String,
-    /// Payment hash for the HTLC for both CKB and BTC.
-    pub payment_hash: Hash256,
-    /// Amount required to pay in Satoshis, including fee
-    #[serde_as(as = "U128Hex")]
-    pub amount_sats: u128,
-    /// Fee in Satoshis
-    #[serde_as(as = "U128Hex")]
-    pub fee_sats: u128,
-    /// Order status
-    pub status: CchOrderStatus,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize)]
-pub struct ReceiveBTCParams {
-    /// Payment request string for the CKB Fiber payee.
-    ///
-    /// The invoice should not be expired soon. The remaining expiry time should be greater than the CCH config
-    /// `min_incoming_invoice_expiry_delta_seconds`.
-    pub fiber_pay_req: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct GetCchOrderParams {
-    /// Payment hash for the HTLC for both CKB and BTC.
-    pub payment_hash: Hash256,
-}
+pub use fiber_json_types::{CchOrderResponse, GetCchOrderParams, ReceiveBTCParams, SendBTCParams};
 
 /// RPC module for cross chain hub demonstration.
 #[rpc(server)]
@@ -187,21 +130,5 @@ impl CchRpcServerImpl {
         })?;
 
         result.map(Into::into).map_err(Into::into)
-    }
-}
-
-impl From<CchOrder> for CchOrderResponse {
-    fn from(value: CchOrder) -> Self {
-        Self {
-            timestamp: value.created_at,
-            expiry_delta_seconds: value.expiry_delta_seconds,
-            wrapped_btc_type_script: value.wrapped_btc_type_script,
-            outgoing_pay_req: value.outgoing_pay_req,
-            incoming_invoice: value.incoming_invoice,
-            payment_hash: value.payment_hash,
-            amount_sats: value.amount_sats,
-            fee_sats: value.fee_sats,
-            status: value.status,
-        }
     }
 }
