+++ crates/fiber-lib/src/rpc/channel.rs
@@ -4,375 +4,26 @@ use crate::fiber::{
         ShutdownCommand, UpdateCommand,
     },
     network::{AcceptChannelCommand, OpenChannelCommand, PendingAcceptChannel},
-    types::{Hash256, Pubkey},
     NetworkActorCommand, NetworkActorMessage,
 };
 use crate::{handle_actor_call, log_and_error};
-use ckb_jsonrpc_types::{EpochNumberWithFraction, Script};
 use ckb_types::{
     core::{EpochNumberWithFraction as EpochNumberWithFractionCore, FeeRate},
-    packed::OutPoint,
     prelude::{IntoTransactionView, Unpack},
-    H256,
 };
-use fiber_types::{
-    AwaitingChannelReadyFlags, AwaitingTxSignaturesFlags, ChannelOpeningStatus,
-    ChannelState as RawChannelState, CloseFlags, CollaboratingFundingTxFlags, EntityHex,
-    NegotiatingFundingFlags, ShuttingDownFlags, SigningCommitmentFlags, TLCId, TlcStatus, U128Hex,
-    U64Hex,
-};
+use fiber_types::{ChannelOpeningStatus, CloseFlags, NegotiatingFundingFlags, TLCId};
 #[cfg(not(target_arch = "wasm32"))]
 use jsonrpsee::proc_macros::rpc;
 
 use jsonrpsee::types::{error::CALL_EXECUTION_FAILED_CODE, ErrorObjectOwned};
 use ractor::{call, ActorRef};
-use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
 use std::cmp::Reverse;
 
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OpenChannelParams {
-    /// The public key of the peer to open a channel with.
-    /// The peer must be connected through the [connect_peer](#peer-connect_peer) rpc first.
-    pub pubkey: Pubkey,
-
-    /// The amount of CKB or UDT to fund the channel with.
-    #[serde_as(as = "U128Hex")]
-    pub funding_amount: u128,
-
-    /// Whether this is a public channel (will be broadcasted to network, and can be used to forward TLCs), an optional parameter, default value is true.
-    pub public: Option<bool>,
-
-    /// Whether this is a one-way channel (will not be broadcasted to network, and can only be used to send payment one way), an optional parameter, default value is false.
-    pub one_way: Option<bool>,
-
-    /// The type script of the UDT to fund the channel with, an optional parameter.
-    pub funding_udt_type_script: Option<Script>,
-
-    /// The script used to receive the channel balance, an optional parameter, default value is the secp256k1_blake160_sighash_all script corresponding to the configured private key.
-    pub shutdown_script: Option<Script>,
-
-    /// The delay time for the commitment transaction, must be an [EpochNumberWithFraction](https://github.com/nervosnetwork/rfcs/blob/master/rfcs/0017-tx-valid-since/e-i-l-encoding.png) in u64 format, an optional parameter, default value is 1 epoch, which is 4 hours.
-    pub commitment_delay_epoch: Option<EpochNumberWithFraction>,
-
-    /// The fee rate for the commitment transaction, an optional parameter.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub commitment_fee_rate: Option<u64>,
-
-    /// The fee rate for the funding transaction, an optional parameter.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub funding_fee_rate: Option<u64>,
-
-    /// The expiry delta to forward a tlc, in milliseconds, default to 4 hours, which is 4 * 60 * 60 * 1000 milliseconds
-    /// Expect it >= 2/3 commitment_delay_epoch.
-    /// This parameter can be updated with rpc `update_channel` later.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub tlc_expiry_delta: Option<u64>,
-
-    /// The minimum value for a TLC our side can send,
-    /// an optional parameter, default is 0, which means we can send any TLC is larger than 0.
-    /// This parameter can be updated with rpc `update_channel` later.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_min_value: Option<u128>,
-
-    /// The fee proportional millionths for a TLC, proportional to the amount of the forwarded tlc.
-    /// The unit is millionths of the amount. default is 1000 which means 0.1%.
-    /// This parameter can be updated with rpc `update_channel` later.
-    /// Not that, we use outbound channel to calculate the fee for TLC forwarding. For example,
-    /// if we have a path A -> B -> C, then the fee B requires for TLC forwarding, is calculated
-    /// the channel configuration of B and C, not A and B.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_fee_proportional_millionths: Option<u128>,
-
-    /// The maximum value in flight for TLCs, an optional parameter.
-    /// This parameter can not be updated after channel is opened.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub max_tlc_value_in_flight: Option<u128>,
-
-    /// The maximum number of TLCs that can be accepted, an optional parameter, default is 125
-    /// This parameter can not be updated after channel is opened.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub max_tlc_number_in_flight: Option<u64>,
-}
-#[derive(Clone, Serialize, Deserialize)]
-pub struct OpenChannelResult {
-    /// The temporary channel ID of the channel being opened
-    pub temporary_channel_id: Hash256,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct AbandonChannelParams {
-    /// The temporary channel ID or real channel ID of the channel being abandoned
-    pub channel_id: Hash256,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct AcceptChannelParams {
-    /// The temporary channel ID of the channel to accept
-    pub temporary_channel_id: Hash256,
-
-    /// The amount of CKB or UDT to fund the channel with
-    #[serde_as(as = "U128Hex")]
-    pub funding_amount: u128,
-
-    /// The script used to receive the channel balance, an optional parameter,
-    /// default value is the secp256k1_blake160_sighash_all script corresponding to the configured private key
-    pub shutdown_script: Option<Script>,
-
-    /// The max tlc sum value in flight for the channel, default is u128::MAX
-    /// This parameter can not be updated after channel is opened.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub max_tlc_value_in_flight: Option<u128>,
-
-    /// The max tlc number in flight send from our side, default is 125
-    /// This parameter can not be updated after channel is opened.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub max_tlc_number_in_flight: Option<u64>,
-
-    /// The minimum value for a TLC our side can send,
-    /// an optional parameter, default is 0, which means we can send any TLC is larger than 0.
-    /// This parameter can be updated with rpc `update_channel` later.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_min_value: Option<u128>,
-
-    /// The fee proportional millionths for a TLC, proportional to the amount of the forwarded tlc.
-    /// The unit is millionths of the amount. default is 1000 which means 0.1%.
-    /// This parameter can be updated with rpc `update_channel` later.
-    /// Not that, we use outbound channel to calculate the fee for TLC forwarding. For example,
-    /// if we have a path A -> B -> C, then the fee B requires for TLC forwarding, is calculated
-    /// the channel configuration of B and C, not A and B.
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_fee_proportional_millionths: Option<u128>,
-
-    /// The expiry delta to forward a tlc, in milliseconds, default to 1 day, which is 24 * 60 * 60 * 1000 milliseconds
-    /// This parameter can be updated with rpc `update_channel` later.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub tlc_expiry_delta: Option<u64>,
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct AcceptChannelResult {
-    /// The final ID of the channel that was accepted, it's different from the temporary channel ID
-    pub channel_id: Hash256,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize)]
-pub struct ListChannelsParams {
-    /// The public key to list channels for.
-    /// An optional parameter, if not provided, all channels will be listed.
-    pub pubkey: Option<Pubkey>,
-    /// Whether to include closed channels in the list, an optional parameter, default value is false
-    pub include_closed: Option<bool>,
-    /// When set to true, only return channels that are still being opened (non-final states:
-    /// negotiating, collaborating on funding tx, signing, awaiting tx signatures, awaiting channel
-    /// ready) as well as channels whose opening attempt failed. Default is false.
-    /// Mutually exclusive with `include_closed`.
-    pub only_pending: Option<bool>,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct ListChannelsResult {
-    /// The list of channels
-    pub channels: Vec<Channel>,
-}
-
-/// The state of a channel
-// `ChannelState` is a copy of `ChannelState` with `#[serde(...)]` attributes for compatibility
-// `bincode` does not support deserialize_identifier
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(
-    rename_all = "SCREAMING_SNAKE_CASE",
-    tag = "state_name",
-    content = "state_flags"
-)]
-pub enum ChannelState {
-    /// We are negotiating the parameters required for the channel prior to funding it.
-    NegotiatingFunding(NegotiatingFundingFlags),
-    /// We're collaborating with the other party on the funding transaction.
-    CollaboratingFundingTx(CollaboratingFundingTxFlags),
-    /// We have collaborated over the funding and are now waiting for CommitmentSigned messages.
-    SigningCommitment(SigningCommitmentFlags),
-    /// We've received and sent `commitment_signed` and are now waiting for both
-    /// party to collaborate on creating a valid funding transaction.
-    AwaitingTxSignatures(AwaitingTxSignaturesFlags),
-    /// We've received/sent `funding_created` and `funding_signed` and are thus now waiting on the
-    /// funding transaction to confirm.
-    AwaitingChannelReady(AwaitingChannelReadyFlags),
-    /// Both we and our counterparty consider the funding transaction confirmed and the channel is
-    /// now operational.
-    ChannelReady,
-    /// We've successfully negotiated a `closing_signed` dance. At this point, the `ChannelManager`
-    ShuttingDown(ShuttingDownFlags),
-    /// This channel is closed.
-    Closed(CloseFlags),
-}
-
-impl ChannelState {
-    /// Returns true if this channel is in a "pending" state — i.e. the channel-opening
-    /// flow has started but has not yet reached `ChannelReady` (and has not been closed
-    /// cooperatively or uncooperatively). This includes `NegotiatingFunding`,
-    /// `CollaboratingFundingTx`, `SigningCommitment`, `AwaitingTxSignatures`, and
-    /// `AwaitingChannelReady`, as well as `Closed(ABANDONED)` / `Closed(FUNDING_ABORTED)`.
-    pub fn is_pending(&self) -> bool {
-        matches!(
-            self,
-            ChannelState::NegotiatingFunding(_)
-                | ChannelState::CollaboratingFundingTx(_)
-                | ChannelState::SigningCommitment(_)
-                | ChannelState::AwaitingTxSignatures(_)
-                | ChannelState::AwaitingChannelReady(_)
-                | ChannelState::Closed(CloseFlags::ABANDONED)
-                | ChannelState::Closed(CloseFlags::FUNDING_ABORTED)
-        )
-    }
-}
-
-impl From<RawChannelState> for ChannelState {
-    fn from(state: RawChannelState) -> Self {
-        match state {
-            RawChannelState::NegotiatingFunding(flags) => ChannelState::NegotiatingFunding(flags),
-            RawChannelState::CollaboratingFundingTx(flags) => {
-                ChannelState::CollaboratingFundingTx(flags)
-            }
-            RawChannelState::SigningCommitment(flags) => ChannelState::SigningCommitment(flags),
-            RawChannelState::AwaitingTxSignatures(flags) => {
-                ChannelState::AwaitingTxSignatures(flags)
-            }
-            RawChannelState::AwaitingChannelReady(flags) => {
-                ChannelState::AwaitingChannelReady(flags)
-            }
-            RawChannelState::ChannelReady => ChannelState::ChannelReady,
-            RawChannelState::ShuttingDown(flags) => ChannelState::ShuttingDown(flags),
-            RawChannelState::Closed(flags) => ChannelState::Closed(flags),
-        }
-    }
-}
-
-/// The channel data structure
-#[serde_as]
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Channel {
-    /// The channel ID
-    pub channel_id: Hash256,
-    /// Whether the channel is public
-    pub is_public: bool,
-    /// Is this channel initially inbound?
-    /// An inbound channel is one where the counterparty is the funder of the channel.
-    pub is_acceptor: bool,
-    /// Is this channel one-way?
-    /// Combines with is_acceptor to determine if the channel able to send payment to the counterparty or not.
-    pub is_one_way: bool,
-    #[serde_as(as = "Option<EntityHex>")]
-    /// The outpoint of the channel
-    pub channel_outpoint: Option<OutPoint>,
-    /// The public key of the channel counterparty.
-    pub pubkey: Pubkey,
-    /// The UDT type script of the channel
-    pub funding_udt_type_script: Option<Script>,
-    /// The state of the channel
-    pub state: ChannelState,
-    /// The local balance of the channel
-    #[serde_as(as = "U128Hex")]
-    pub local_balance: u128,
-    /// The offered balance of the channel
-    #[serde_as(as = "U128Hex")]
-    pub offered_tlc_balance: u128,
-    /// The remote balance of the channel
-    #[serde_as(as = "U128Hex")]
-    pub remote_balance: u128,
-    /// The received balance of the channel
-    #[serde_as(as = "U128Hex")]
-    pub received_tlc_balance: u128,
-    /// The list of pending tlcs
-    pub pending_tlcs: Vec<Htlc>,
-    /// The hash of the latest commitment transaction
-    pub latest_commitment_transaction_hash: Option<H256>,
-    /// The time the channel was created at, in milliseconds from UNIX epoch
-    #[serde_as(as = "U64Hex")]
-    pub created_at: u64,
-    /// Whether the channel is enabled
-    pub enabled: bool,
-    /// The expiry delta to forward a tlc, in milliseconds, default to 1 day, which is 24 * 60 * 60 * 1000 milliseconds
-    /// This parameter can be updated with rpc `update_channel` later.
-    #[serde_as(as = "U64Hex")]
-    pub tlc_expiry_delta: u64,
-    /// The fee proportional millionths for a TLC, proportional to the amount of the forwarded tlc.
-    /// The unit is millionths of the amount. default is 1000 which means 0.1%.
-    /// This parameter can be updated with rpc `update_channel` later.
-    /// Not that, we use outbound channel to calculate the fee for TLC forwarding. For example,
-    /// if we have a path A -> B -> C, then the fee B requires for TLC forwarding, is calculated
-    /// the channel configuration of B and C, not A and B.
-    #[serde_as(as = "U128Hex")]
-    pub tlc_fee_proportional_millionths: u128,
-    /// The hash of the shutdown transaction
-    pub shutdown_transaction_hash: Option<H256>,
-    /// Human-readable reason why the channel opening failed.
-    /// Only present when the channel is in a failed state (e.g. abandoned or funding aborted).
-    pub failure_detail: Option<String>,
-}
-
-/// The htlc data structure
-#[serde_as]
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Htlc {
-    /// The id of the htlc
-    #[serde_as(as = "U64Hex")]
-    pub id: u64,
-    /// The amount of the htlc
-    #[serde_as(as = "U128Hex")]
-    pub amount: u128,
-    /// The payment hash of the htlc
-    pub payment_hash: Hash256,
-    /// The expiry of the htlc
-    #[serde_as(as = "U64Hex")]
-    pub expiry: u64,
-    /// If this HTLC is involved in a forwarding operation, this field indicates the forwarding channel.
-    /// For an outbound htlc, it is the inbound channel. For an inbound htlc, it is the outbound channel.
-    pub forwarding_channel_id: Option<Hash256>,
-    /// If this HTLC is involved in a forwarding operation, this field indicates the forwarding tlc id.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub forwarding_tlc_id: Option<u64>,
-    /// The status of the htlc
-    pub status: TlcStatus,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ShutdownChannelParams {
-    /// The channel ID of the channel to shut down
-    pub channel_id: Hash256,
-    /// The script used to receive the channel balance, only support secp256k1_blake160_sighash_all script for now
-    /// default is `default_funding_lock_script` in `CkbConfig`
-    pub close_script: Option<Script>,
-    /// The fee rate for the closing transaction, the fee will be deducted from the closing initiator's channel balance
-    /// default is 1000 shannons/KW
-    #[serde_as(as = "Option<U64Hex>")]
-    pub fee_rate: Option<u64>,
-    /// Whether to force the channel to close, when set to false, `close_script` and `fee_rate` should be set, default is false.
-    /// When set to true, `close_script` and `fee_rate` will be ignored and will use the default value when opening the channel.
-    pub force: Option<bool>,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct UpdateChannelParams {
-    /// The channel ID of the channel to update
-    pub channel_id: Hash256,
-    /// Whether the channel is enabled
-    pub enabled: Option<bool>,
-    /// The expiry delta for the TLC locktime
-    #[serde_as(as = "Option<U64Hex>")]
-    pub tlc_expiry_delta: Option<u64>,
-    /// The minimum value for a TLC
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_minimum_value: Option<u128>,
-    /// The fee proportional millionths for a TLC
-    #[serde_as(as = "Option<U128Hex>")]
-    pub tlc_fee_proportional_millionths: Option<u128>,
-}
+pub use fiber_json_types::{
+    AbandonChannelParams, AcceptChannelParams, AcceptChannelResult, Channel, ChannelState, Htlc,
+    ListChannelsParams, ListChannelsResult, OpenChannelParams, OpenChannelResult,
+    ShutdownChannelParams, UpdateChannelParams,
+};
 
 /// RPC module for channel management.
 #[cfg(not(target_arch = "wasm32"))]
