+++ crates/fiber-lib/src/rpc/invoice.rs
@@ -6,210 +6,25 @@
 use crate::fiber::config::{MAX_PAYMENT_TLC_EXPIRY_LIMIT, MIN_TLC_EXPIRY_DELTA};
 use crate::fiber::{NetworkActorCommand, NetworkActorMessage};
 use crate::invoice::{
-    Attribute as InternalAttribute, CkbInvoice as InternalCkbInvoice, CkbInvoiceStatus, CkbScript,
-    Currency, InvoiceBuilder, InvoiceData as InternalInvoiceData, InvoiceSignature, InvoiceStore,
+    CkbInvoice as InternalCkbInvoice, CkbInvoiceStatus, Currency, InvoiceBuilder, InvoiceStore,
 };
 use crate::{gen_rand_sha256_hash, handle_actor_call, log_and_error, FiberConfig};
-use fiber_types::{duration_hex, FeatureVector, Hash256, HashAlgorithm, Privkey, U128Hex, U64Hex};
+use fiber_types::{FeatureVector, Privkey};
 
-use ckb_jsonrpc_types::Script;
 #[cfg(not(target_arch = "wasm32"))]
 use jsonrpsee::proc_macros::rpc;
 use jsonrpsee::types::{error::CALL_EXECUTION_FAILED_CODE, ErrorObjectOwned};
 use ractor::{call, ActorRef};
 use rand::Rng;
 use secp256k1::{PublicKey, SecretKey, SECP256K1};
-use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
 use std::time::Duration;
 use tentacle::secio::SecioKeyPair;
 
-/// The attributes of the invoice
-#[serde_as]
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Attribute {
-    #[serde(with = "U64Hex")]
-    /// This attribute is deprecated since v0.6.0, The final tlc time out, in milliseconds
-    FinalHtlcTimeout(u64),
-    #[serde(with = "U64Hex")]
-    /// The final tlc minimum expiry delta, in milliseconds, default is 1 day
-    FinalHtlcMinimumExpiryDelta(u64),
-    #[serde(with = "duration_hex")]
-    /// The expiry time of the invoice, in seconds
-    ExpiryTime(Duration),
-    /// The description of the invoice
-    Description(String),
-    /// The fallback address of the invoice
-    FallbackAddr(String),
-    /// The udt type script of the invoice
-    UdtScript(CkbScript),
-    /// The payee public key of the invoice
-    PayeePublicKey(PublicKey),
-    /// The hash algorithm of the invoice
-    HashAlgorithm(HashAlgorithm),
-    /// The feature flags of the invoice
-    Feature(Vec<String>),
-    /// The payment secret of the invoice
-    PaymentSecret(Hash256),
-}
-
-/// The metadata of the invoice
-#[serde_as]
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct InvoiceData {
-    /// The timestamp of the invoice
-    #[serde_as(as = "U128Hex")]
-    pub timestamp: u128,
-    /// The payment hash of the invoice
-    pub payment_hash: Hash256,
-    /// The attributes of the invoice, e.g. description, expiry time, etc.
-    pub attrs: Vec<Attribute>,
-}
-
-/// Represents a syntactically and semantically correct lightning BOLT11 invoice
-///
-/// There are three ways to construct a `CkbInvoice`:
-///  1. using [`CkbInvoiceBuilder`]
-///  2. using `str::parse::<CkbInvoice>(&str)` (see [`CkbInvoice::from_str`])
-///
-#[serde_as]
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct CkbInvoice {
-    /// The currency of the invoice
-    pub currency: Currency,
-    #[serde_as(as = "Option<U128Hex>")]
-    /// The amount of the invoice
-    pub amount: Option<u128>,
-    /// The signature of the invoice
-    pub signature: Option<InvoiceSignature>,
-    /// The invoice data, including the payment hash, timestamp and other attributes
-    pub data: InvoiceData,
-}
-
-impl From<InternalAttribute> for Attribute {
-    fn from(attr: InternalAttribute) -> Self {
-        match attr {
-            InternalAttribute::FinalHtlcTimeout(timeout) => Attribute::FinalHtlcTimeout(timeout),
-            InternalAttribute::FinalHtlcMinimumExpiryDelta(delta) => {
-                Attribute::FinalHtlcMinimumExpiryDelta(delta)
-            }
-            InternalAttribute::ExpiryTime(duration) => Attribute::ExpiryTime(duration),
-            InternalAttribute::Description(desc) => Attribute::Description(desc),
-            InternalAttribute::FallbackAddr(addr) => Attribute::FallbackAddr(addr),
-            InternalAttribute::UdtScript(script) => Attribute::UdtScript(script),
-            InternalAttribute::PayeePublicKey(pubkey) => Attribute::PayeePublicKey(pubkey),
-            InternalAttribute::HashAlgorithm(alg) => Attribute::HashAlgorithm(alg),
-            InternalAttribute::Feature(feature) => {
-                Attribute::Feature(feature.enabled_features_names())
-            }
-            InternalAttribute::PaymentSecret(secret) => Attribute::PaymentSecret(secret),
-        }
-    }
-}
-
-impl From<InternalInvoiceData> for InvoiceData {
-    fn from(data: InternalInvoiceData) -> Self {
-        InvoiceData {
-            timestamp: data.timestamp,
-            payment_hash: data.payment_hash,
-            attrs: data.attrs.into_iter().map(|a| a.into()).collect(),
-        }
-    }
-}
-
-impl From<InternalCkbInvoice> for CkbInvoice {
-    fn from(inv: InternalCkbInvoice) -> Self {
-        CkbInvoice {
-            currency: inv.currency,
-            amount: inv.amount,
-            signature: inv.signature,
-            data: inv.data.into(),
-        }
-    }
-}
-
-/// The parameter struct for generating a new invoice.
-#[serde_as]
-#[derive(Serialize, Deserialize, Default, Clone)]
-pub struct NewInvoiceParams {
-    /// The amount of the invoice.
-    #[serde_as(as = "U128Hex")]
-    pub amount: u128,
-    /// The description of the invoice.
-    pub description: Option<String>,
-    /// The currency of the invoice.
-    pub currency: Currency,
-    /// The preimage to settle an incoming TLC payable to this invoice. If preimage is set, hash must be absent. If both preimage and hash are absent, a random preimage is generated.
-    pub payment_preimage: Option<Hash256>,
-    /// The hash of the preimage. If hash is set, preimage must be absent. This condition indicates a 'hold invoice' for which the tlc must be accepted and held until the preimage becomes known.
-    pub payment_hash: Option<Hash256>,
-    /// The expiry time of the invoice, in seconds.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub expiry: Option<u64>,
-    /// The fallback address of the invoice.
-    pub fallback_address: Option<String>,
-    /// The final HTLC timeout of the invoice, in milliseconds.
-    /// Minimal value is 16 hours, and maximal value is 14 days.
-    #[serde_as(as = "Option<U64Hex>")]
-    pub final_expiry_delta: Option<u64>,
-    /// The UDT type script of the invoice.
-    pub udt_type_script: Option<Script>,
-    /// The hash algorithm of the invoice.
-    pub hash_algorithm: Option<HashAlgorithm>,
-    /// Whether allow payment to use MPP
-    pub allow_mpp: Option<bool>,
-    /// Whether allow payment to use trampoline routing
-    pub allow_trampoline_routing: Option<bool>,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct InvoiceResult {
-    /// The encoded invoice address.
-    pub invoice_address: String,
-    /// The invoice.
-    pub invoice: CkbInvoice,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct ParseInvoiceParams {
-    /// The encoded invoice address.
-    pub invoice: String,
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct ParseInvoiceResult {
-    /// The invoice.
-    pub invoice: CkbInvoice,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct InvoiceParams {
-    /// The payment hash of the invoice.
-    pub payment_hash: Hash256,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct SettleInvoiceParams {
-    /// The payment hash of the invoice.
-    pub payment_hash: Hash256,
-    /// The payment preimage of the invoice.
-    pub payment_preimage: Hash256,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct SettleInvoiceResult {}
-
-/// The status of the invoice.
-#[derive(Clone, Serialize, Deserialize)]
-pub struct GetInvoiceResult {
-    /// The encoded invoice address.
-    pub invoice_address: String,
-    /// The invoice.
-    pub invoice: CkbInvoice,
-    /// The invoice status
-    pub status: CkbInvoiceStatus,
-}
+pub use fiber_json_types::{
+    Attribute, CkbInvoice, GetInvoiceResult, InvoiceData, InvoiceParams, InvoiceResult,
+    NewInvoiceParams, ParseInvoiceParams, ParseInvoiceResult, SettleInvoiceParams,
+    SettleInvoiceResult,
+};
 
 /// RPC module for invoice management.
 #[cfg(not(target_arch = "wasm32"))]
