+++ crates/fiber-lib/src/rpc/peer.rs
@@ -1,5 +1,4 @@
 use crate::fiber::network::PeerDisconnectReason;
-use crate::fiber::types::Pubkey;
 use crate::fiber::{NetworkActorCommand, NetworkActorMessage};
 use crate::log_and_error;
 #[cfg(not(target_arch = "wasm32"))]
@@ -9,47 +8,9 @@ use jsonrpsee::types::ErrorObjectOwned;
 
 use ractor::call;
 use ractor::ActorRef;
-use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
 pub use tentacle::multiaddr::MultiAddr;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ConnectPeerParams {
-    /// The address of the peer to connect to.
-    /// Either `address` or `pubkey` must be provided.
-    pub address: Option<MultiAddr>,
-    /// The public key of the peer to connect to.
-    /// The node resolves the address from locally synced graph data.
-    pub pubkey: Option<Pubkey>,
-    /// Whether to save the peer address to the peer store.
-    pub save: Option<bool>,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DisconnectPeerParams {
-    /// The public key of the peer to disconnect.
-    pub pubkey: Pubkey,
-}
-
-/// The information about a peer connected to the node.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PeerInfo {
-    /// The identity public key of the peer.
-    pub pubkey: Pubkey,
-
-    /// The multi-address associated with the connecting peer.
-    /// Note: this is only the address which used for connecting to the peer, not all addresses of the peer.
-    /// The `graph_nodes` in Graph rpc module will return all addresses of the peer.
-    pub address: MultiAddr,
-}
-
-/// The result of the `list_peers` RPC method.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ListPeersResult {
-    /// A list of connected peers.
-    pub peers: Vec<PeerInfo>,
-}
+pub use fiber_json_types::{ConnectPeerParams, DisconnectPeerParams, ListPeersResult, PeerInfo};
 
 /// RPC module for peer management.
 #[cfg(not(target_arch = "wasm32"))]
